Removed, organized files, and updated file paths
