--- client/src/components/Authentication/LocalSignUpForm.jsx
@@ -1,18 +1,0 @@
-import React, { useState } from "react";
-import Input from "../UI/input";
-import SubmitButton from "../UI/submitButton";
-
-function LocalSignUpForm() {
-    const [username, setUsername] = useState("");
-    const [password, setPassword] = useState("");
-
-    return (
-        <div>
-            <form action="/create" method="POST">
-
-            </form>
-        </div>
-    );
-}
-
-export default LocalSignUpForm;
