--- client/src/components/INTERNAL_Experiments/2_successEnterPage.jsx
@@ -1,11 +1,0 @@
-import React from "react";
-
-function SuccessSignInPage() {
-    return (
-        <div>
-            <h3>YAY! SIGN IN SUCCESSFULLY</h3>
-        </div>
-    );
-}
-
-export default SuccessSignInPage;
