+++ client/src/pages/InsertPage/auth_InsertPage/auth_InsertPage.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import Input from "../UI/input";
+import Input from "../../../components/Input/input";
 
 function SecretEnterPage() {
     const [companyName, setCompanyName] = useState("");
